ticker_etl/pyproject.toml
[project]
name = "ticker_etl"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "pathlib>=1.0.1",
    "polygon-api-client>=1.15.3",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "pytz>=2025.2",
    "ruff>=0.12.9",
]

[build-system]
requires = ["uv_build>=0.8.0,<0.9.0"]
build-backend = "uv_build"
